Tool-level support for using postgresql.
Various modifications to allow migrations to run (mostly) on a local
postgresql database, and to allow test data to be loaded into it.

Updated build.gradle to include a postgresql JDBC dependency, and to
produce a correct JDBC URL for a postgresql database (with sensible
defaults for local development).

Updated migrations to specify that constraint-related migrations should
be run in any non-sqlite database, not just mssql, and to include a
postgresql-friendly version of the two migrations that do an INSERT.

Added a perl script to convert insertBaseData.sql to postgresql-friendly
behavior, and modified a few lines of insertBaseData.sql to have the
correct (camel) case for the table being inserted.