Initial postgresql support in search code.
Updated AnetObjectEngine and DaoUtils to support switch-based instead of
boolean-based multiple database support, and added a bunch of code around
Reports searching to allow the SqliteReportSearcher to work for either sqlite
or pg, depending on arguments passed in at construction time.

This included some refactoring that it seemed to make sense to apply to
MssqlReportSearcher as well, so I did that.