Updated Sqlite*Searcher classes to also support postgresql.
Since postgresql and SQLite are largely compatible, at least as an
interim measure we can use the SQLite searchers for postgresql as well,
as long as they handle case-sensitive column names correctly. Updated
the SQL generation code to do that, and to impose ordering on Person
searches in the outer query as well as the id-finding subquery.