NCI-Agency/anet#333: Sort full-text search results on search rank
If we're doing a full-text search, and there is no explicit sort order,
sort first on the search pseudo-rank.
If the full-text search also contains a LIKE predicate, give that match
the highest possible score (1000 points).
Refactor/reorganize the code to make the classes more similar/consistent
in the way the SQL clauses are constructed.