Updated TaskDao and Poam-to-task migrations to support postgresql.
Updated the requoting script to handle the new identifiers for Tasks,
and ran it on TaskDao. Updated an old foreign-key-adding migration
to run on postgresql as well as mssql, which allows all the table-
and column-renaming migrations to run on postgresql as well, and
updated one of those migrations to have a valid rollback, since it
was needed during development.