NCI-Agency/anet#416: Fix cursor behaviour in person firstname field
The issue was that whenever something was typed in the firstname field,
the cursor would go immediately to the end of the input. This was
happening because the event target value was being changed by the JS
code in order to remove the eventual comma added through the lastname
field. I fixed this by changing the way the comma is removed.