NCI-Agency/anet#375: Update dependencies for React 16
This updates the version of the react modules in order to be compatible
with React 16. It also removes some dependencies like alloy-editor and
react-injectables as they are not React 16 compatible yet. We will have
to figure out later how to deal with the modules which are not React 16
compatible yet.