NCI-Agency/anet#375: Remove advanced search button on search page
Using the advanced search button on the search page was not working
properly: when clicking the button and clicking save search afterwards
we were getting the error "TypeError: advancedSearch.filters". We
removed the button as we also don't see why it should be used, people
can just use the search bar.