NCI-Agency/anet#588: Remove the progress bar for now
This is needed as a quick fix because right now because of nprogress we
get an error when trying to load one of the reports (The error is:
Unable to get property 'className' of undefined or null reference) and
this results in a blank page.
We will replace the component later by a different one.