Fix "Map container is already initialized"
When there are multiple maps on a page (because there are multiple report collections on a page), Leaflet may throw this error because all div's have the same id.
Pass a new, optional mapId prop to fix this case.
Also, only try to add/update layers after the state has been updated.