NCI-Agency/anet#657: Move show/hide loading from LoaderHOC
This moves the showLoading and hideLoading dispatch call from the render
method of the LoaderHOC component to the insight components. This way we
avoid the warning "Cannot update during an existing state transition
(such as within render or another component's constructor)." We keep the
LoaderHOC in order to display the loading on the bar charts too.