GH-551 Add dynamic topbar height calculation
In order to use the fixed header and overflow-y for both the sidebar and
the main content area, the height of the top bar should be known. Since
the top bar height is variable, the height needs to be calculated
dynamically. This can then be added to the top padding of the content
area, so that the height of the sidebar and content area are know.