NCI-Agency/anet#890: Do no extra search on click on search nav
This makes sure that when clicking on an item of the search navigation
we no longer perform a new search but just go to the anchor of the
selected search result type.

The search navigation now also uses scrollspy and the everything option
has been removed from the navigation as the page always contains all
results.

This also makes sure that clicking an anchor link no longer results in
the page being reloaded.