NCI-Agency/anet#822: Upgrade GraphQL Java dependencies
As of GraphQL 0.9.8, enums are no longer of type <EnumType>Input but just <EnumType>.
This also makes pageNum and pageSize required for search queries (since they are int primitives);
add them where they are missing.
For the home tiles we are not actually interested in the reports themselves, just the totalCount,
so retrieve at most one report.