NCI-Agency/anet#1098: Dispatch clearSearchQuery only on nav click
The searchQuery Redux state needs to be cleared when clicking on a nav
item. When dispatching the clear action in the constructor of the page
we had problems because the page was finishing it's lifecycle first
using the old searchQuery state, which resulted in GraphQL errors when
the searchQuery state was about people rank and the page query was for
reports. This is why I moved the dispatch action call to the onClick
event in the Nav component.