NCI-Agency/anet#1175: Fetch data on ReportVisualisation component mount
This makes sure that when the user goes from an insight to a report and
than uses the back button to go back to the insight , the insight no
longer shows no data. This was happening because the data was not
fetched again if the search state did not change.