NCI-Agency/anet#255: Reset pageNum on fetchResults in certain situations
This resets the first page to 0 when changing the search terms and when
changing filters in the attendees overlay.

We want to make sure that changing the search terms does a new search
and starts with the first page of results. Also changing tabs now
results in resetting the pagination to show the first page.