NCI-Agency/anet#255: Do not filter out the selected attendees in the overlay
This makes sure we no longer filter out the selected attendees in the
attendees overlay. This because filtering them client side results in
strange problems with the list of attendees in the overlay in
combination with the pagination: some pages would display less items.

We choose this instead of fixing filtering server side as this also has
the least impact on the user: otherwise the pagination would change
while more and more items are being added.