Replaced buildup phase with converge mechanism
Some stories will influence a number (e.g. createPerson will increase the
number of persons). Depending on that number, a mean and a standard deviation
the story will be executed or skipped with a certain chance.
The idea is to have stories and 'counteracting' stories (e.g. createPerson
and deletePerson) to converge to a number that varies around the mean with a
certain standard deviation.

implements nci-agency/anet#379