NCI-Agency/anet#256: Add approval actions for submit, reject and publish actions
This makes sure that not only approval actions related to an approval
step are saved as an approval action but also submit, reject by admin
and publish actions. This way we can see the whole workflow of a report
in time.

When a report is in the approved or published state we now display the
whole workflow, not only the one related to approval steps.