NCI-Agency/anet#1252: Add reports search which does not exclude states
When the report search is part of the automatic report publication
action we need to be able to retrieve all approved reports. We therefore
made it possible to tell the report search to not to exclude certain
report states: by marking a reports search as being a system search (a
search action not being performed by a user but by the system as part of
an automated action like a search needed for a worker). This search does
not exclude draft, approved and rejected reports.