NCI-Agency/anet#1252: Sort approved reports on dashboard
This makes sure that the approved reports on the dashboard of an admin
user are sorted ascending op updateDate. We actually want them to be
sorted on approval timestamp, but directly after the approval action is
added the report is updated too, so sorting on updated_at should result
in the same order.