NCI-Agency/anet#1252: Add submit and publish actions for existing reports
This adds a migration script to add report actions for the submit and
publish of a report. For the submit we use the timestamp of the first
report action of the report (this one would be a reject or approve
action) and than substitute a second from it to make sure the submit
action was first.

For publish we use the release date of reports having the sate RELEASED.