NCI-Agency/anet#1252: Rename released to publish
This renames the released term to published. The releasedAt property of
a report and other related variables stay the same as they are relating
to the publication or cancellation date.

We used to have a published reports subitem under My Reports displaying
released and cancelled reports. We now added a Cancelled reports subitem
and the Published reports now only display the published one.

Note that when we refer to the release of a report from now on we refer
to the publication or the cancellation.