NCI-Agency/anet#255: Remove extra results reset step
This makes sure that when the search term in the advanced multi select
widget changes, we no longer first change the results state to empty
(which results in an extra load of the widget and showing an empty page
of results). We now reset the results just after fetching the data, and
update it with the fetched data. This way we avoid the following when
typing in the input widget: lets say we are in "All advisors", we have
typed "ja" so far and we add another letter to "jac" => The results
flicker from 2 results - to empty - to 2 results.