NCI-Agency/anet#1367: Verify end of tour date not in past
When creating a new user as an admin or, when a new user status
logs in for the first time, then verify that the end of tour
date is not in the past.

When creating a new user, use newUserForm test to validate new user.
Then check if the end of tour date is in the future.

Set value must first clear existing input before running test