NCI-Agency/anet#1326: Update display of advanced select widget filters
This makes sure that when there is only one filter we no longer display
it. It also makes sure that in a small overlay (like in the advanced
search filters, where we have an overlay in an overlay) the filters are
displayed on the tip of the table of results, as a dropdown.