NCI-Agency/anet#1326: Make it possible to search without query params
For the advanced-select-widget it should be possible to retrieve all
items of a certain type, without any query params (filters). This has
been made possible now.

But for the search results and insights page we do not want people to be
able to get all items by not using any search criteria. We therefore
show an error message when one tries to search without search criteria.