NCI-Agency/anet#1326: Remove recent filters from report fields widgets
This removes for now the recents filters from the report fields having
an advanced-select-widget. The recents are already available as
shortcuts in the right column. We will probably readd them when they
will also use a search query and therefore will also be filterable by
text filled in the input field.