NCI-Agency/anet#1633: Fix push history state
When using hashSpy in scroll link nav item the history gets
rewritten. This causes the back button to stop working.

In order to fix this we do not set hashSpy, and implement
the function pushHash ourselves.

react-scroll has a PR ready to fix this, however the PR has not
been merged since december.
https://github.com/fisshy/react-scroll/pull/359/commits/101b0e203168e9827737581e82450c0db678f690