NCI-Agency/anet#1668: Fix validation message not showing
When the model constructor called set state a check is performed
that removes keys with a null value. Yup schema relies on all
keys that need to be checked to be present in the initial values.
By removing the conditional validations are also run for fields
with a default value of null.