NCI-Agency/anet#GH-1811: Add new line plugin
The new line plugin adds paragraphs to the pasting of content in
IE 11. The plugin is not activated in browsers that properly
handle rich text paste events.

The order of the plugins is relevant for linkify plugin to work.
First add linkify, then the new line plugin in the editor.