NCI-Agency/anet#1723: No longer display no reports found before search
This makes sure that when loadign an insight page or a saved search, the
report collection is not being rendered before the reports are fetched.
Otherwise we would first get the message "no reports found" and
afterwards the eventual reports would be displayed. This is misleading,
especially if a search is a bit slow.