NCI-Agency/anet#1723: Add reportsFilter to the ReportCollection
This makes it possible again to add a reports filter to the
ReportCollection container. This time we provide the component as a
property and the ReportCollection component doesn't contain logic about
when to display the filter or not. Which filter is to be used and when
is decided in parent components.