Remove return null on page loading
Removed returning null on page render when the page is loading,
introduced in commit 293ceb9. This was probably added in order to avoid
displaying things like no reports found during data fetching. But we
fixed that at the level of displaying the reports collection instead of
on the general page level. Changing between renderPage and null
penending on whether the page was loading or not was resulting in the
page subcomponents being mountedseveral times, which had side efects if
such a component was doing data fetching during componentDidMount (data
fetching was done several times, for instance in the
ReportCollectionContainer).