NCI-Agency/anet#1722: Fix getting the future engagements for the worker
The FutureEngagementsWorker was also retrieving past engagements which
were not first upcoming engagements. This because the check for pending
approval reports for which the planning approval action didn't take
place yet was not correct: it was not checking specifically if the
report was attached to a planning type of approval step. This has been
fixed now.