NCI-Agency/anet#1722: Fix e2e tests when run after the integration tests
The report publish chain tests were considering the first pending
approval and approved report. But if the tests were run after the server
side integration tests (like Travis does) the first report was no longer
the report supposed to be used. We now changed the test to search the
report by url instead of by position in the list.