NCI-Agency/anet#1810: Use blueprint popover for advanaced select
The popover from blueprint can be customized a bit regarding it's
position: with tricks we can now align the popver under both the field
label and the field widget, even though it is related only to the field
widget.

But, this is not responsive, when the window is too narrow the popover
is not displayed completely.