NCI-Agency/anet#1810: Replace blueprintjs popover with bootstrap one
This replaces the blueprintjs popover with the bootstrap one in the
advanced select widget. This because the blueprintjs popover had the
side effect that the ContainerDimensions was giving randomly different
widths back, and because of that the filters were sometimes displayed as
nav and sometimes as dropdown, for the same browser window width.