Fix relative date filters in saved searches
When a saved search was containing a relative date filter the end date
used to be set to the date when the saved search was created. This was
resulting in the search no longer being relative to the current date.

We now set the end date server side, when creating the sql statement,
which fixes the problem.