Try to avoid deadlocks when checking authorization
While running parallel simulator scenarios, we would sometimes get:
  com.microsoft.sqlserver.jdbc.SQLServerException: Transaction (Process ID #) was deadlocked on lock resources with another process and has been chosen as the deadlock victim. Rerun the transaction.
and the exception would always be in the `checkReportAuthorization` query.
Probable cause is that that query tried to access the reports table while it was still locked (because of report insertion/update).
Avoid using the reports table altogether, by just checking the author from the report object.