NCI-Agency/anet#2352: Do not update the form values to make custom fields undefined
In order to make sure that the invisible custom fields are not being
saved in the JSON, we used to make their corresponding value within the
values object undefined. Because of that after an autosave we were
sometimes getting a warning about changing a component from uncontrolled
to controlled.

We now no longer change the values variable. This has also as
side-effect that if one has filled in a custom field and that one has
become invisible afterwards, the field still has the filled-in value
after it becomes visible again.