Improve performance of Formik text fields in report form
This makes sure that for text fields we only call validate onBlur. This
way, for text fields formik no longer does also validation when typing a
character.

Unfortunatelly this only works when we set the validateOnChange to false
and then use setFieldValue(fieldname, value, true) for the fields which
should use validation on change.