Make sure input fields are also validated
In order to improve the performance of filling in input fields we were
no longer doing validation on change for these ones, but only on blur.
We now also added validation onChange, but a debounced validation. Also,
by using now the validateField method we make sure that the validation
of input fields is only doing validation of the field schema, not of the
form schema.