NCI-Agency/anet#2682: Change squash of report actions when pending approval
For the report workflow, when a report was pending approval, we used to
only add report actions for the approval steps for which we had no
action. Therefore, after report resubmission (after a reject action), we
could no longer see pending actions in the end.

We now only display the report actions since the last submit, including
the pending ones. This way, for a report pending approval the user sees
the current approval workflow since the last submit.

Updated the test for this as well.