NCI-Agency/anet#2579: Update AdvancedSelect component
When typing one letter in the input field, the component was doing first
a debounced fetch for ean empty string and afterwards one for the typed
letter. Fixed that to only do one debounced fetch.

Also made sure that if one of the fetch results end after a next one,
nothing is being done with it's results.

Simplified the code a bit by removing not longer needed handlers.