NCI-Agency/anet#2579: Do not fetch results when closing popover
Because state update is not always batched, we were getting an extra
fetch results when closing the popover. We were also getting too many
rerenders in that case. We fixed it by doing the state update as part of
an effect in this case, this way we make sure it is being batched.