Store the avatar as binary data in the database
Previously, it was stored as base64-encoded data in a binary field.
It was also using Liquibase's default blob type for PostgreSQL (oid), which is cumbersome to use
(and would make it harder to migrate from MS SQL Server to PostgreSQL).
Note: the database migration temporarily needs about twice the storage space for all avatars.

The contract between client and server (that we offer base64-encoded output and expect base64-encoded input) is now in the GraphQL methods for the avatar in Person.java.