NCI-Agency/anet#2905: Extract AddAssessmentModal component
We extracted the AddAssessmentModal component from the task show page.
We also made sure that when the modal is closed after a new monthly
assessment has been added the current task is refetched, in order
to get the latest notes, both on the page as in the side pannel for
notes.