Update documentation about nodejs and yarn setup
We no longer install nodejs separately, it is being added as a
dependency when ./gradlew run does the client build step.
Also yarn install is not being needed as a separate step, these
dependencies are also added as a part of the step mentioned above.
What is needed is to make sure that NODEJS_HOME and JARN_HOME are
being set (udi==sing the dependencies from .gradle) and added
to the path.