NCI-Agency/anet#2926: Update yup schema for task assessment
The task assessment is using a custom fields definition. Custom
fields schema should be a key-object type of value, the key being
also the prefix to use in forms for all fields related to the task
assessments. For this purpose we changed the schema shape by adding
the taskAssessment key.