NCI-Agency/anet#2926: Code refactoring ReadonlyCustomFields
ReadonlyCustomFields actually only need the formik related property
called values. For readonly fields we don't need to do setFieldValue,
setFieldTouched.

In order to make it easier to understand the code, we have removed
the formikProps from readonly custom fields and only provide the
values property directly.

This is also a first step in trying to make the readonly custom fields
no longer depend on Formik in order to reduce complexity.