NCI-Agency/anet#2926: Fix display of last assessments
The last assessment fields of type enum were not being displayed
because ReadonlyEnumField is a Formik FastField and it's humanValue
property is a function, in which case the ReadonlyField component
uses information which is only available if the field is being a
child of a Formik component.
To fix this we've added the Formik component as parent for the last
assessment readonly fields.

Another solution we might consider later is to no longer make the
readonly custom fields depend on Formik by calculating
the value to be used for the humanValue ourselves.