NCI-Agency/anet#2926: Code refactoring add assessment conditions
Refactored the conditions in which the add period assessment is being displayed:
- it is related to the given period: whether for the period is allowed to
add assessments (allowNewAssessments property of the assessmentPeriod object)
- it is related to whether this should be done for the current assessments results
component (see the canAddAssessment property of the assessment results component).

For subentities we set this one to false. For the entity itself this
property is true when the user has the proper rights to be able to add
assessments.