NCI-Agency/anet#2926: Update period related logic for getting assessments
First of all we now include the start and end date ranges when filtering
the assessments.

Moved the calculation of the period when retrieving the last assessment
to the caller of getLastAssessment, the method should just retrieve the
last assessment made in the given period, it is on the caller to
decide on the period depending for which purpose this is being used.

When retrieving the last assessment, set as end of the period the end of
the month, as otherwise when adding just one month to the end of February
we would not get the end of March.