NCI-Agency/anet#2954: Core refactoring assessments terminology
This renames the assessments made on tasks when working on reports
(in the context of the report form) to measurements. It also makes
the difference in terminology more clear in the AssessmentResultsTable.
Measurements are assessments made on an entity in relation with another entity:
so measurements of an engagement's related tasks while filling the
engagement form, or measurements of an engagement's attendees while
filling the engagement form.
Further we call assessments the assessments made about an entity
related to a given period of time: like month assessments of a task
or of a person.