NCI-Agency/anet#2926: Refactor AggregationWidget number aggregation
This makes sure that when using a number type of aggregation we
only consider number values or the values which can be transformed
to a number. We also moved away the transformation from the
aggregation calculations, we make sure we transform the values
before providing them to the aggregation functions.