NCI-Agency/anet#2954: Add task assessments custom field
This adds an assessments customField to task, which makes it
possible to define different types of assessments for a task.
The field is an array_of_objects type of field, and each object
will be the definition of a new assessments.

If these assessments are being defined for a task, we use them,
otherwise we use the old way of defining the instant assessments
and the periodic monthly assessments.