NCI-Agency/anet#2977: Add general assessments definitions to tasks
We now added the possibility to add assessments definitions for a
task on a more general level: per task level. In anet.yml these
assessments definitions can be configured using the assessments
property for fields.task.topLevel and fields.task.subLevel.
When a task also has assessments defintions at an instance level
(by filling the task's assessments custom field), we make the
union of the general and the instance specific defintions and the
instance specific definition has priority.