NCI-Agency/anet#2977: Simplify retrieval of assessments config
Because the report tasks on the report form and on the show page were
not always Task instances the code for retrieving the assessments
configuration for tasks was quite complicated, making it possible to
be run on both Task instances as just objects.
We now made sure that the report tasks on these pages are always
task instances.