NCI-Agency/anet#2977: Fix AddAssessmentModal for multiple periods
When using the AddAssessmentModal for more than one periods it was
always opening the modal for the last period, and therefore also
using that period's details when saving the assessment => all
assessments were belonging to the same period.

Fixed this by moving the button to open the modal in the modal
component itself.