NCI-Agency/anet#2977: Display all assessments for a period
This makes sure we no longer display only the last assessment for
a period but all assessments: more users can make an assessment for
a period.
This also makes sure that each user can only make one assignment,
by only displaying the add assessment button when the user didn't
already make an assessment for the given period.

The laout of the periodic has also changed: we now use a panel to
make it easier to distinuish between different assessments and also
in order to make it possible to display the author and an eventual
edit/delete icon.