NCI-Agency/anet#2954: Fix adding report task assessments for recents
When adding a task from the recents list the task assessments were
not being updated, this has been fixed now.
We now also make sure that when adding items from FieldShortcuts
we always return an instance of the given object type, just as the
advanced select widget also does. This has been done as right now
the field shortcuts are used in combination with the advanced select
widget, and we would expect from both the same return type.