NCI-Agency/anet#2971: Make default aggregation depend on the widget
The default aggregation used to be determined based on the field type.
The aggregation widget was also determined based on the field type.
But that way, if a user configured the widget but didn't configure
the aggregation the default aggregation for the field type would
have been used and might have ended up in errors.