Fix CustomFields enumHumanValue
Make sure the function doesn't give an error when the fieldValue is
not among the choices.
Saw one context in which an enum value was suddenly the string null
(can't reproduce it any more) and we were getting an error. Added
the fix to make sure that if that happens again we don't get an
error.