NCI-Agency/anet#2971: Improve pie aggregation widget
Always display the number of entities in the middle and make a pie
slice for each used value. For enumset types of fields we might have
more selected values per entity so the number displayed in the middle
of the pie no longer corresponds with the count of the different values.