NCI-Agency/anet#2971: Fix ReportMapWidget error
When being on the rollup page and clicking on the email rollup button
we were getting an error that the map was already initialized. This
was because of the way we were determining the mapId.
A page contains several map widgets whe used for the report
statistics. We therefore had to make sure each map on a mage has a
unique id.