NCI-Agency/anet#2971: Change no reports message in ReportStatistics
The message "no reports found" was not always clear enough for
ReportStatistics. In the following context it needed extra details:
- I create a report without an engagement date (a draft report)
- I go to My Reports, draft reports
- I see in the summary view a report (the draft I've just created)
- I switch to the statistics tab
- I see now "no reports found"
This would be strange to the users, why are there reports in the
summary tab but not in the statistics tab? This happens because for
the statistics we only retrieve the reports for the past 3 months,
so in this context no reports are retrieved for statistics.

In order to make it clear to the user what happens, we made the
no reports found message more specific for ReportStatistics.