NCI-Agency/anet#2971: Improve ReportsByTask aggregation widget
This makes sure that the ReportsByTask aggregation widget no longer
displays a chart with a bar for no effort when there are no reports
to aggregate the data of. Instead it just displays the "no data"
message, as the reports by task insights also do.

Also did a small code style refactoring in the ReportsByTasksWidget.