NCI-Agency/anet#3081:Add template with readonly blocks to reportText
This makes sure that when creating a report, the reportText field
contains a template text. Some of the heading blocks in the template
are readonly and can't be removed.
Note: it only works now on add, after save, on edit, the headings
are no longer readonly (still to be implemented).