NCI-Agency/anet#3081: Update richTextEditor mandatory h1
This makes sure that if the richTextEditor contains h1 tags marked
with the mandatory class, they can be edited but the tags itself
can not be deleted. Also, when the content is saved, the resulting
html also still has the mandatory class for the h1 tag (this was not
the case before).