NCI-Agency/anet#3081: Make placeholder configurable for mandatory elements
This makes sure that when specifying a default template for a
richTextEditor type of field, we can also define a placeholder for
the html elements having the class mandatory. This placeholder
is being used to prefill the element if this one becomes empty.
Both the class attribute and the placeholder attribute of an HTML
tag of type h1, h2, h3 or p are being persisted in the database. This
in order to be able to still know the mandatory elements and their
eventual placeholders when editing an object with rich text content.