Make more link popover visible above the top bar
Since fullcalendar version 5, the more link popover is a fixed position
element. Because of this it is being displayed behind the top bar.
This makes sure that when the calendar is being displayed on pages other
than insights the popover is being displayed above the top bar.

Note: on insights pages the more link popover is behind quite a lot
of elements.