Fix prettier adding trailing commas before eslint removes them
when we run yarn run lint-fix, prettier formats with trailing commas
And eslint removes those commas. Making build system fail for a second
And cause a rebuild + ide and git warn about the problems in the code
for nothing. If we don't already allow trailing commas, we should also
prevent prettier to format that way