NCI-Agency/anet#3221: Add reportActions.planned column
For planned engagements that were automatically approved (because there was no planning workflow),
the FutureEngagementWorker had no way of finding these, and thus they were never set back to draft status when the time came.
We solve this by indicating if an (automatic) ReportAction was for a planned engagement.

Add tests for the FutureEngagementWorker and the ReportPublicationWorker.