NCI-Agency/anet#3208: Fix getting responsibleTasks
Fixes part of commit 5c8df12ae0fcd1f4b3cdedb159bb7d0cbac4b3f4,
which would always get the responsibleTasks for the current user doing the request,
instead of those belonging to the Person object.
Move responsibleTasks to Position, as that is the proper place for it,
and reflect this change in the JavaScript code requesting them.