NCI-Agency/anet#3327: Refactor recent users & activities
Replace user's domainUsername in the admin view with a LinkTo to the person.
Only show the 10 most recent activities for a user (instead of 100).
Format the time according to the setting in the dictionary.
Also use specific toastId's so we don't show the same message more than once.
Suppress warnings about serialVersionUID.