NCI-Agency/anet#1344: Add My Subscriptions to My Work menu
Move the My Subscriptions and My Subscription Updates tables to a new My Subscriptions page under the My Work menu.
Also make sure the My Subscription Updates table is updated when unsubscribing through the My Subscriptions table (and vice versa).
When a user deletes the last entry on the last page of the table, display "nothing to show…" so they might move to the previous page.