AB#300: Make sure buttons are disabled when they should be
And also make sure the tooltip (in the `title` prop) for buttons is shown.
With disabled buttons, we need to wrap them in an extra element, as disabled buttons do not respond
to mouse events, and in that case, we also need to make sure that they do not receive any mouse
events.