AB#219: Add header to ignore background polling activity
Split into request filter and response filter, and use the request
filter for logging user activity, and the response filter for caching.
Check for the "x-activity: ignore" header to ignore polling requests.
Don't log personally identifiable information; log what comes out of the
security principal's `toString()`, which for a Person is its uuid.
Refactor the code a bit for readability.