AB#895 Correct avatar in organizational chart
Use the correct data URL for avatars in the organizational chart;
updates commit 650647b4bf205e7386e8a3f7bedefcabdc50cd19
Modify the webpack config so you can decide on a case-by-case basis to
inline an image or not.
Make sure the organization icons and the default avatar in the chart are
inlined (so exported SVG diagrams are self-contained).