AB#1005 Improve MIME type checks
If the MIME type suggested by the browser when creating an attachment
differs from the detected MIME type based on the uploaded attachment contents,
update the MIME type if the detected MIME type is also allowed.
This helps with container formats like HEIC/HEIF, where the browser may
not be able to accurately guess the actual format based on just the
filename extension.